requirements.txt
pillow==9.0.0
python-telegram-bot==22.0
python-dotenv==1.0.0

